fix: adapt to tauri-plugin-notifications refacto
