+++ example/matrix-svelte-client/src-tauri/gen/android/app/proguard-rules.pro
@@ -21,9 +21,6 @@
 #-renamesourcefileattribute SourceFile
 -keep, includedescriptorclasses class org.rustls.platformverifier.** { *; }
 
-# The silent-push handler is only reachable through the SILENT_PUSH_HANDLER
-# manifest meta-data + Class.forName, and the JNI bridge only through the
-# handler — R8 can't see either reference, so without these keeps release
-# builds strip them and lose the killed-state (cold) push path.
--keep class com.matrix.svelte.client.DemoSilentPushHandler { <init>(); }
--keep class com.matrix.svelte.client.SilentPushBridge { *; }
+# Silent pushes: the JNI bridge (SilentPushNative) now lives in the
+# notifications plugin, whose consumer-rules.pro carries the keep — no
+# app-side rules needed.
