+++ example/matrix-svelte-client/src-tauri/gen/apple/project.yml
@@ -121,7 +121,7 @@ targets:
   # Notification Service Extension: iOS launches it for every push carrying
   # `mutable-content: 1` (any app state, incl. force-quit). Its principal class
   # subclasses the plugin's TauriNotificationService, which calls the Rust
-  # handler `src/ios_push.rs` exports from the same libapp.a the app links.
+  # handler `src/push_handler.rs` exports from the same libapp.a the app links.
   matrix-svelte-client_NSE:
     type: app-extension
     platform: iOS
@@ -162,7 +162,7 @@ targets:
         LIBRARY_SEARCH_PATHS[arch=arm64]: $(inherited) $(PROJECT_DIR)/Externals/arm64/$(CONFIGURATION) $(SDKROOT)/usr/lib/swift $(TOOLCHAIN_DIR)/usr/lib/swift/$(PLATFORM_NAME) $(TOOLCHAIN_DIR)/usr/lib/swift-5.0/$(PLATFORM_NAME)
         # libapp.a is a static archive: force the dlsym'd Rust entry points in,
         # or the linker dead-strips them (an "undefined symbol" error here
-        # means ios_silent_push_handler! wasn't invoked in the Rust crate).
+        # means silent_push_handler! was not invoked in the Rust crate).
         # Do NOT add -ObjC: it would force-load every ObjC class in libapp.a.
         OTHER_LDFLAGS: $(inherited) -Wl,-u,_tauri_notifications_process_silent_push -Wl,-u,_tauri_notifications_silent_push_free
     dependencies:
