--- example/matrix-svelte-client/src-tauri/src/ios_push.rs
@@ -1,102 +1,0 @@
-//! iOS silent-push decoding for the Notification Service Extension (NSE).
-//!
-//! iOS hands every push carrying `mutable-content: 1` (the shape the Matrix
-//! pusher registers — with a fallback `alert`) to the app's NSE target, in
-//! **every** app state including force-quit, with ~30 s to rewrite the
-//! content. The NSE links the same staticlib as the app (linking it does not
-//! start Tauri, mirroring how the Android FCM service loads the same `.so`)
-//! and resolves the C entry point generated by
-//! [`tauri_plugin_notifications::ios_silent_push_handler!`] below via `dlsym`.
-//! There is no Tauri `AppHandle` in that process, so we return the content as
-//! [`NotificationData`] and let the extension render it.
-//!
-//! `data_dir` is the App Group container shared with the main app (see the
-//! `TauriNotificationsAppGroup` key in the NSE's Info.plist): the plugin
-//! stores the Matrix store, salt file, and — through the shared keychain
-//! access group — the session there when `ios_app_group` is configured, so
-//! this process can decrypt the pushed event exactly like the Android cold
-//! path does.
-//!
-//! If this handler returns `None` (or fails in any way), the extension shows
-//! the push's own fallback `alert` (`loc-key: SINGLE_UNREAD`) instead.
-
-use std::collections::HashMap;
-
-use tauri_plugin_notifications::{NotificationData, NotificationMessage};
-
-/// Must match `ios_app_group` in `tauri.conf.json` and the App Group in both
-/// targets' entitlements.
-const APP_GROUP: &str = "group.com.matrix.svelte.client";
-
-/// Decode a Matrix push (`room_id` / `event_id` custom keys) into the
-/// notification to display. Runs inside the NSE process.
-// The by-value map is the `SilentPushHandler` contract (the handler owns the data).
-#[allow(clippy::needless_pass_by_value)]
-fn handle_silent_push(data_dir: &str, data: HashMap<String, String>) -> Option<NotificationData> {
-    // The NSE process never runs the plugin `setup`, so replay the keyring
-    // init here (idempotent), scoped to the shared access group the app
-    // saved the session into.
-    if let Err(e) = tauri_plugin_matrix_svelte::init_keyring_store(Some(APP_GROUP)) {
-        tracing::error!("NSE: couldn't init keyring store: {e}");
-        return None;
-    }
-
-    // No room/event → not a Matrix message push; keep the fallback alert.
-    let room_id = data.get("room_id")?.clone();
-    let event_id = data.get("event_id")?.clone();
-
-    // A per-push runtime is fine here: the multi-process NSE mode restores a
-    // fresh client on every call anyway (no cached client/tasks to keep alive,
-    // unlike the Android cold path's process-wide runtime).
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .ok()?;
-    let content = runtime.block_on(crate::push_shared::fetch_notification_event(
-        data_dir.to_owned(),
-        room_id.clone(),
-        event_id.clone(),
-    ));
-
-    // No `.id(...)`: the extension cannot change the identifier APNs assigned.
-    // `group` (→ `threadIdentifier`) stacks the room's notifications together
-    // instead. The deep link and the Matrix ids ride in `extra`, whose string
-    // values surface in the `notificationClicked` event's `data` when tapped —
-    // the frontend routes `deepLink` through `handleMatrixUri`.
-    //
-    // The `message`/`conversation_title`/`group_conversation` fields mirror the
-    // Android MessagingStyle payload (`android_push.rs`): the NSE renders them
-    // as a communication notification. DMs draw the sender's avatar in place
-    // of the app icon; group rooms brand as the room instead — room name as
-    // the group title, room avatar (when it has one) as the icon. Requires the
-    // communication-notifications entitlement on the app (see
-    // `gen/apple/project.yml`).
-    let mut message = NotificationMessage::new(content.body.clone())
-        .sender(content.sender.clone())
-        .person_key(content.sender);
-    if let Some(avatar) = content.sender_avatar {
-        message = message.avatar_bytes(avatar);
-    }
-    let mut builder = NotificationData::builder()
-        .title(content.summary)
-        .body(content.body)
-        .group(room_id.as_str())
-        .summary(content.room_display_name.clone())
-        .message(message)
-        .conversation_title(content.room_display_name)
-        .extra(
-            "deepLink",
-            crate::push_shared::matrix_uri(&room_id, &event_id),
-        )
-        .extra("room_id", room_id)
-        .extra("event_id", event_id);
-    if !content.is_dm {
-        builder = builder.group_conversation();
-        if let Some(avatar) = content.room_avatar {
-            builder = builder.conversation_avatar_bytes(avatar);
-        }
-    }
-    Some(builder.build())
-}
-
-tauri_plugin_notifications::ios_silent_push_handler!(handle_silent_push);
