+++ example/matrix-svelte-client/src-tauri/src/lib.rs
@@ -257,12 +257,11 @@ pub fn run() {
                     })
                 });
             // Silent (data-only) pushes are handled natively: the notifications
-            // plugin's FCM service dispatches every one — warm or killed — to
-            // `DemoSilentPushHandler` (manifest meta-data), which fetches the
-            // content through the JNI entry in `android_push` and posts the
-            // notification itself. No Rust-side `on_silent_push` handler: it
-            // would only run if the native handler failed, and keeping a second
-            // builder path in lockstep proved error-prone.
+            // plugin's FCM service loads this library (`SILENT_PUSH_LIB`
+            // manifest meta-data) and dispatches every one — warm or killed —
+            // to the handler `push_handler` registers with
+            // `silent_push_handler!`, then posts the returned notification
+            // itself.
             #[cfg(target_os = "android")]
             {
                 // Dismiss a room's notification (and its stored conversation)
@@ -276,7 +275,7 @@ pub fn run() {
                         tracing::error!("room-read event with unexpected payload: {}", event.payload());
                         return;
                     };
-                    let id = android_push::notification_id_for(&room_id);
+                    let id = push_handler::notification_id_for(&room_id);
                     // A single-id list: an empty list would mean "cancel all".
                     if let Err(e) = dismiss_handle.notifications().remove_active(vec![id]) {
                         tracing::error!("failed to dismiss notification for read room {room_id}: {e}");
@@ -434,25 +433,22 @@ pub extern "system" fn Java_com_matrix_svelte_client_MainActivity_initNdkContext
     _class: jni::objects::JObject,
     context: jni::objects::JObject,
 ) {
-    // Shared with the cold-push JNI entry so `ndk_context` is initialized at most
-    // once per process. The FCM service runs in the same process and Android may
-    // reuse a push process to launch this activity; using one guard for both
-    // entry points avoids `ndk_context`'s double-init panic. See `android_push`.
-    android_push::ensure_ndk_context(&mut env, &context);
+    // Shared with the silent-push init hook so `ndk_context` is initialized at
+    // most once per process. The FCM service runs in the same process and
+    // Android may reuse a push process to launch this activity; using one
+    // guard for both entry points avoids `ndk_context`'s double-init panic.
+    // See `push_handler`.
+    push_handler::ensure_ndk_context(&mut env, &context);
 }
 
-// Android-only background (killed-state) silent-push handling: a JNI entry the
-// FCM service calls when there is no Tauri runtime. See the module docs.
-#[cfg(target_os = "android")]
-mod android_push;
+// The one silent-push handler for both mobile platforms (iOS NSE dlsym entry +
+// Android FCM JNI entry, both generated by `silent_push_handler!`), plus the
+// Android init hook replaying process-wide setup on cold starts. See the
+// module docs.
+#[cfg(any(target_os = "android", target_os = "ios"))]
+mod push_handler;
 
-// iOS-only silent-push decoding: the C entry point the app's Notification
-// Service Extension resolves via dlsym. See the module docs.
-#[cfg(target_os = "ios")]
-mod ios_push;
-
-// Platform-agnostic fetch/format helpers shared by all silent-push entry
-// points: Android warm + JNI killed paths, and the iOS NSE path.
+// Platform-agnostic fetch/format helpers shared by the silent-push handler.
 #[cfg(any(target_os = "android", target_os = "ios"))]
 mod push_shared;
